+++ examples/vite-react/src/App.tsx
@@ -4,17 +4,17 @@ import "./App.css";
 import {
   KitEventType,
   SwkAppDarkTheme,
-} from "@creit-tech/stellar-wallets-kit/types";
-import { StellarWalletsKit } from "@creit-tech/stellar-wallets-kit/sdk";
+} from "@creit.tech/stellar-wallets-kit/types";
+import { StellarWalletsKit } from "@creit.tech/stellar-wallets-kit/sdk";
 import {
   activeAddress,
   activeModule,
-} from "@creit-tech/stellar-wallets-kit/state";
-import { ButtonMode } from "@creit-tech/stellar-wallets-kit/components";
-import { defaultModules } from "@creit-tech/stellar-wallets-kit/modules/utils";
-import { TrezorModule } from "@creit-tech/stellar-wallets-kit/modules/trezor";
-import { LedgerModule } from "@creit-tech/stellar-wallets-kit/modules/ledger";
-import { WalletConnectModule } from "@creit-tech/stellar-wallets-kit/modules/wallet-connect";
+} from "@creit.tech/stellar-wallets-kit/state";
+import { ButtonMode } from "@creit.tech/stellar-wallets-kit/components";
+import { defaultModules } from "@creit.tech/stellar-wallets-kit/modules/utils";
+import { TrezorModule } from "@creit.tech/stellar-wallets-kit/modules/trezor";
+import { LedgerModule } from "@creit.tech/stellar-wallets-kit/modules/ledger";
+import { WalletConnectModule } from "@creit.tech/stellar-wallets-kit/modules/wallet-connect";
 
 import {
   Account,
@@ -66,14 +66,14 @@ export class App extends Component<any, any> {
      */
     StellarWalletsKit.on(KitEventType.STATE_UPDATED, (event) => {
       console.log(`Address updated:`, event.payload.address);
-      this.setState({address: event.payload.address});
+      this.setState({ address: event.payload.address });
     });
     StellarWalletsKit.on(KitEventType.WALLET_SELECTED, (event) => {
       console.log(`Wallet ID:`, event.payload.id);
-      this.setState({moduleId: event.payload.id});
+      this.setState({ moduleId: event.payload.id });
     });
     StellarWalletsKit.on(KitEventType.DISCONNECT, () => {
-      this.setState({address: undefined, productName: undefined});
+      this.setState({ address: undefined, productName: undefined });
     });
     StellarWalletsKit.createButton(document.querySelector("#button")!, {
       mode: ButtonMode.free,
@@ -82,7 +82,7 @@ export class App extends Component<any, any> {
 
   async authModal(): Promise<void> {
     try {
-      const {address} = await StellarWalletsKit.authModal();
+      const { address } = await StellarWalletsKit.authModal();
       console.log(`Address fetched:`, address);
     } catch (e) {
       console.error(e);
@@ -94,12 +94,18 @@ export class App extends Component<any, any> {
   }
 
   async signTransaction(): Promise<void> {
-    const {address} = await StellarWalletsKit.getAddress();
+    const { address } = await StellarWalletsKit.getAddress();
     console.log("StellarWalletsKit::getAddress", address);
-    const tx = new TransactionBuilder(new Account('GD47IF5ARQLIVXUAX6EFN7TELAU6DUYQHPM7DNNLZ7VSR5TQSW6LQV74', "-1"), {
-      networkPassphrase: Networks.PUBLIC,
-      fee: "0",
-    })
+    const tx = new TransactionBuilder(
+      new Account(
+        "GD47IF5ARQLIVXUAX6EFN7TELAU6DUYQHPM7DNNLZ7VSR5TQSW6LQV74",
+        "-1",
+      ),
+      {
+        networkPassphrase: Networks.PUBLIC,
+        fee: "0",
+      },
+    )
       .setTimeout(0)
       .addOperation(
         Operation.manageData({
@@ -109,10 +115,10 @@ export class App extends Component<any, any> {
       )
       .build();
 
-    console.log(tx.toXDR());
+    console.log(tx.toXdr());
 
-    const {signedTxXdr} = await StellarWalletsKit.signTransaction(
-      tx.toXDR(),
+    const { signedTxXdr } = await StellarWalletsKit.signTransaction(
+      tx.toXdr(),
       {
         networkPassphrase: Networks.PUBLIC,
         address,
@@ -123,11 +129,11 @@ export class App extends Component<any, any> {
   }
 
   async signAuthEntry(): Promise<void> {
-    const {address} = await StellarWalletsKit.getAddress();
+    const { address } = await StellarWalletsKit.getAddress();
     const xdr =
       "AAAACXrDOZdUTjF10ma9AiQ5sizbFlCMARY/JuXLKj4QRal5Ueb3t2qeufIDkl6TAAAAAAAAAAHD5Dhm6FraoWtNw3xmsftfw43aav9gLsi5kDYD1ccr/gAAAApzdGFydF9nYW1lAAAAAAAFAAAAAy2DQRwAAAASAAAAAAAAAACO+drsns+C8ivJ7BbEvGPuuaf+RI7JYRYQh3tTDoG6yAAAABIAAAAAAAAAADzwlU9pvQCCZwaS876OkOohieXRjEidV8RoVpxgVhODAAAACgAAAAAAAAAAAAAAAAAPQkAAAAAKAAAAAAAAAAAAAAAAAA9CQAAAAAEAAAAAAAAAAQ711IO2j2xojpimQQ1dzE4A9Kskd2MeHXPKwLGFTKYYAAAACnN0YXJ0X2dhbWUAAAAAAAYAAAASAAAAAcPkOGboWtqha03DfGax+1/Djdpq/2AuyLmQNgPVxyv+AAAAAy2DQRwAAAASAAAAAAAAAACO+drsns+C8ivJ7BbEvGPuuaf+RI7JYRYQh3tTDoG6yAAAABIAAAAAAAAAADzwlU9pvQCCZwaS876OkOohieXRjEidV8RoVpxgVhODAAAACgAAAAAAAAAAAAAAAAAPQkAAAAAKAAAAAAAAAAAAAAAAAA9CQAAAAAA=";
 
-    const {signedAuthEntry} = await StellarWalletsKit.signAuthEntry(xdr, {
+    const { signedAuthEntry } = await StellarWalletsKit.signAuthEntry(xdr, {
       networkPassphrase: Networks.PUBLIC,
       address,
     });
@@ -142,39 +148,39 @@ export class App extends Component<any, any> {
         <div className="card">
           <section>
             <div>
-              <button onClick={ () => this.authModal() }>
+              <button onClick={() => this.authModal()}>
                 Connect Wallet
               </button>
 
-              <button onClick={ () => this.disconnect() }>
+              <button onClick={() => this.disconnect()}>
                 Disconnect
               </button>
 
-              <button onClick={ () => this.signTransaction() }>
+              <button onClick={() => this.signTransaction()}>
                 Sign transaction
               </button>
 
-              <button onClick={ () => this.signAuthEntry() }>
+              <button onClick={() => this.signAuthEntry()}>
                 Sign auth Entry
               </button>
 
-              <button onClick={ () => StellarWalletsKit.profileModal() }>
+              <button onClick={() => StellarWalletsKit.profileModal()}>
                 Profile Modal
               </button>
             </div>
 
             <div>
               <p>
-                Your selected wallet is: <br/> { this.state.moduleId }
+                Your selected wallet is: <br /> {this.state.moduleId}
               </p>
               <p>
-                Your account is: <br/> { this.state.address &&
-                `${ this.state.address.slice(0, 4) }....${
-                  this.state.address.slice(-6)
-                }` }
+                Your account is: <br /> {this.state.address &&
+                  `${this.state.address.slice(0, 4)}....${
+                    this.state.address.slice(-6)
+                  }`}
               </p>
 
-              <div style={ {marginBottom: "3rem"} }></div>
+              <div style={{ marginBottom: "3rem" }}></div>
 
               <div>
                 <p>